[selectors] :focus-visible should stop matching after blur https://bugs.webkit.org/show_bug.cgi?id=233688
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Add new test to check that :focus-visible stop matching after blur.

* web-platform-tests/css/selectors/focus-visible-023-expected.txt: Added.
* web-platform-tests/css/selectors/focus-visible-023.html: Added.

Source/WebCore:

Move PseudoClassChangeInvalidation for :focus-visible together with the one for :focus,
as we need to keep both flags in sync when computing the invalidations
(it's not possible to have an element that matches :focus-visible but not :focus,
and that was happening before this change).

Test: imported/w3c/web-platform-tests/css/selectors/focus-visible-023.html

* dom/Element.cpp:
(WebCore::Element::setFocus): Add PseudoClassChangeInvalidation for :focus-visible change,
and put it in the same scope as the flags changes.
(WebCore::Element::setHasFocusVisible): Remove PseudoClassChangeInvalidation from this method.
* page/EventHandler.cpp:
(WebCore::EventHandler::internalKeyEvent): Use PseudoClassChangeInvalidation for :focus-visible change.



Canonical link: https://commits.webkit.org/244761@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286415 268f45cc-cd09-0410-ab3c-d52691b4dbfc