[Web Animations] Add a way to run scripted animations https://bugs.webkit.org/show_bug.cgi?id=233869 rdar://85983542
Reviewed by Dean Jackson.

Source/WebCore:

Tests: webanimations/custom-effect/custom-effect.html
       webanimations/custom-effect/document-timeline-animate.html

This patch adds two new Web-exposed features to allow authors to write callback-based animations
leveraging the full power of the Web Animations model.

First, we add a new AnimationEffect subclass which wraps a callback to be executed on every tick
of the associated animation's timeline: CustomEffect. It can be constructed by providing a JS function
as its first parameter, and either a number or a dictionary to provide timing properties for the effect.
This is similar to how KeyframeEffect will accept a keyframes object and then timing data. The callback
is provided the effect progress as its sole parameter. Web authors can use this new interface as follows:

    const animation = new Animation;
    animation.effect = new CustomEffect(progress => { … }, 1000);
    animation.play();

Second, to make starting callback-based animations more straightforward and in a very similar fashion
to how keyframe animations can be initiated on an element using Element.animate(), we introduce a new
animate() method on DocumentTimeline, allowing the previous exmaple to be written as:

    const animation = document.timeline.animate(progress => { … }, 1000);

This simple approach allows Web authors to move past the simple use of requestAnimationFrame() and harness
the Web Animations timing model which will let them pause and resume animations, seek them, control their
playback rate, apply easing, respond to the "finished" promise, etc.

The code itself is very simple, CustomEffect is simply an AnimationEffect subclass that indicates that it's
always interested in scheduling updates while active using the ticksContinouslyWhileActive() method. This
means that all the scheduling logic contained in WebAnimation and AnimationEffect applies, allowing callbacks
to not be fired when the animation is paused or when the page is suspended, or fired sparingly when a steps()
timing function is specified.

* CMakeLists.txt:
* DerivedSources-input.xcfilelist:
* DerivedSources-output.xcfilelist:
* DerivedSources.make:
* Headers.cmake:
* Sources.txt:
* WebCore.xcodeproj/project.pbxproj:
* animation/AnimationEffect.h:
(WebCore::AnimationEffect::isCustomEffect const):
* animation/CustomAnimationOptions.h: Added.
* animation/CustomAnimationOptions.idl: Added.
* animation/CustomEffect.cpp: Added.
(WebCore::CustomEffect::create):
(WebCore::CustomEffect::CustomEffect):
(WebCore::CustomEffect::animationDidTick):
* animation/CustomEffect.h: Added.
(WebCore::CustomEffect::~CustomEffect):
* animation/CustomEffect.idl: Added.
* animation/CustomEffectCallback.h: Added.
* animation/CustomEffectCallback.idl: Added.
* animation/DocumentTimeline.cpp:
(WebCore::DocumentTimeline::animate):
* animation/DocumentTimeline.h:
* animation/DocumentTimeline.idl:
* bindings/js/WebCoreBuiltinNames.h:

Source/WTF:

Add a new experimental feature for the CustomEffect interface.

* Scripts/Preferences/WebPreferencesExperimental.yaml:

LayoutTests:

Add tests for the new CustomEffect interface and the document.timeline.animate() method.
These are written using WPT libraries such that they may be upstreamed to the WPT repository
in the future if and when the CustomEffect interface is standardized.

* platform/win/TestExpectations:
* webanimations/custom-effect/custom-effect-expected.txt: Added.
* webanimations/custom-effect/custom-effect.html: Added.
* webanimations/custom-effect/document-timeline-animate-expected.txt: Added.
* webanimations/custom-effect/document-timeline-animate.html: Added.


Canonical link: https://commits.webkit.org/244885@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286555 268f45cc-cd09-0410-ab3c-d52691b4dbfc