[css-flexbox] Account for captions when flexing tables with specified sizes https://bugs.webkit.org/show_bug.cgi?id=233814
Reviewed by Darin Adler.

Source/WebCore:

Flexing tables is complex because of the specifics of the table layout algorithms. There are
two interrelated issues that were causing some failures in flexbox WPT tests.

The first one is that tables interpret overriding sizes (flexed sizes) as the sizes of rows + captions.
However the specified height of a table only accounts for the heights of the rows, not the captions. That's
why when setting the flexed height of a table we must add up the specified height of the table and the height
of the captions. The table algorithm will properly substract the captions' size in order to compute the
available height for rows.

The second issue is that the table layout algorithm adds the size of the bottom border at the very end
of its execution after performing all the computations and it does it unconditionally. The thing is that
the flexbox code already takes into account the borders and paddings so we basically have to substract that
border when setting the height of the table, because the table layout will add it later.

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::computeMainAxisExtentForChild):
* rendering/RenderTable.cpp:
(WebCore::RenderTable::computeCaptionsLogicalHeight const): New method refactored from current code.
(WebCore::RenderTable::layout):
* rendering/RenderTable.h: Expose computeCaptionsLogicalHeight.

LayoutTests:

* TestExpectations: Unskipped two tests that are now passing


Canonical link: https://commits.webkit.org/244918@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286593 268f45cc-cd09-0410-ab3c-d52691b4dbfc