Let RemoteAudioMediaStreamTrackRendererInternalUnitManager::Unit switch to VPIO unit if VPIO is running https://bugs.webkit.org/show_bug.cgi?id=233316
Reviewed by Eric Carlson.

Source/WebCore:

Add support for an external provider of speaker data to the audio capture unit.
We use this in WebKit to render the MediaStreamTracks of the process doing capture through VPIO when running.
In case the audio capture unit speaker format is not matching the producer format, we have to reconfigure and align with the producer format.
In case, audio is being rendered by the audio capture unit but capture is stopped, we continue running the audio capture unit.
This allows to mute/unmute microphone without any audio glitch.
We make sure to stop the audio capture unit when it is no longer used for rendering or capture.

In case the unit is suspended and we restart an audio source, we will unsuspend the unit, which will unmute sources.
We do not want to start audio sources within the start of an audio source.
For that reason, we unmute the sources asynchronously.

Manually tested.

* WebCore.xcodeproj/project.pbxproj:
* platform/mediastream/mac/BaseAudioSharedUnit.cpp:
* platform/mediastream/mac/BaseAudioSharedUnit.h:
* platform/mediastream/mac/CoreAudioCaptureSource.cpp:
* platform/mediastream/mac/CoreAudioCaptureSource.h:

Source/WebKit:

We track which process is the last to start audio capture.
In that case, we register its RemoteAudioMediaStreamTrackRendererInternalUnitManager::Unit to the capture unit.
This allows RemoteAudioMediaStreamTrackRendererInternalUnitManager::Unit to switch between RemoteIO if capture unit does not run
or the capture unit if running.
This allows to remove the x5 gain workaround.

* GPUProcess/GPUConnectionToWebProcess.cpp:
* GPUProcess/GPUConnectionToWebProcess.h:
* GPUProcess/GPUProcess.cpp:
* GPUProcess/GPUProcess.h:
* GPUProcess/webrtc/RemoteAudioMediaStreamTrackRendererInternalUnitManager.cpp:
* GPUProcess/webrtc/RemoteAudioMediaStreamTrackRendererInternalUnitManager.h:
* UIProcess/Cocoa/UserMediaCaptureManagerProxy.cpp:
* UIProcess/Cocoa/UserMediaCaptureManagerProxy.h:



Canonical link: https://commits.webkit.org/244963@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286650 268f45cc-cd09-0410-ab3c-d52691b4dbfc