Add a way to write a `SharedBuffer` to the Pasteboard https://bugs.webkit.org/show_bug.cgi?id=234065
Reviewed by Wenson Hsieh.

In order to match the WK1 implementation of `-pasteFont:` <https://webkit.org/b/191379>, we
need to be able to write a `SharedBuffer` (with an associated type) to the font pasteboard.

Create a `struct PasteboardBuffer` to wrap the `SharedBuffer` and `String type` (and a
`String contentOrigin` on Cocoa platforms, which is used to decide whether to show a "Paste"
callout to the user as a gate on crossorigin content sharing) in a single object.

Though it isn't used in this patch, `-pasteFont:` <https://webkit.org/b/191379> will use
this to write RTF data (which contains the necessary font data).

No change in behavior.

* platform/Pasteboard.h:
* platform/Pasteboard.cpp:
* platform/libwpe/PasteboardLibWPE.cpp:
(WebCore::Pasteboard::write):
* platform/gtk/PasteboardGtk.cpp:
(WebCore::Pasteboard::write):
* platform/ios/PasteboardIOS.mm:
(WebCore::Pasteboard::write):
* platform/mac/PasteboardMac.mm:
(WebCore::Pasteboard::write):
* platform/win/PasteboardWin.cpp:
(WebCore::Pasteboard::write):



Canonical link: https://commits.webkit.org/245094@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286861 268f45cc-cd09-0410-ab3c-d52691b4dbfc