WebCore::createDOMException() should abort early if termination is pending. https://bugs.webkit.org/show_bug.cgi?id=234190
Reviewed by Darin Adler.

Attempting to create Error objects may re-enter the VM, which we should not do
when termination is pending.

This issue manifested as an ASSERT failure, and was discovered while running
http/wpt/fetch/ layout tests with a Debug build on an M1 Mac.  It also manifested
on some testing bots.

* bindings/js/JSDOMExceptionHandling.cpp:
(WebCore::createDOMException):



Canonical link: https://commits.webkit.org/245138@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286912 268f45cc-cd09-0410-ab3c-d52691b4dbfc