Regression(r283565) Unable to report private messages as Spam in Twitter app https://bugs.webkit.org/show_bug.cgi?id=234253 <rdar://86043227>
Reviewed by Alex Christensen.

Source/WebCore:

Add macro for new linkedOnOrAfter check for sending the Authorization header
on same origin redirects.

* platform/cocoa/VersionChecks.h:

Source/WebKit:

In r283565, we aligned WebKit's behavior with the Fetch specification and with
Gecko / Blink by keeping the Authorization header on same origin redirects.

However, when reporting a private message in the twitter app, it does a
same-origin redirect from `https://twitter.com/account/authenticate_web_view?...`
to `https://twitter.com/account/authenticate_web_view?...` and the HTTP server
reponds with a 401/Unauthorized when we send the Authorization on the post-redirect
request. As far as I can tell, our behavior is correct here and I suspect this is an
issue with the twitter server. As a result, I am simply gating the new behavior
behind a linked-on-or-after check to give twitter a chance to address the issue on
their end.

* NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::NetworkDataTaskCocoa::willPerformHTTPRedirection):

Source/WTF:

Add some new iOS / macOS versions.

* wtf/spi/darwin/dyldSPI.h:



Canonical link: https://commits.webkit.org/245193@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286971 268f45cc-cd09-0410-ab3c-d52691b4dbfc