[GTK] Fix ANGLE compilation https://bugs.webkit.org/show_bug.cgi?id=233327
Patch by Alejandro G. Castro <alex@igalia.com> on 2021-12-13
Reviewed by Kenneth Russell.

Source/ThirdParty/ANGLE:

There is a define with a typo in the name and activating EGL and
X11 does not work at the same time.

* PlatformGTK.cmake:

Source/WebCore:

No tests required.

Fixed the issues to make the engine work with -DUSE_ANGLE_WEBGL=ON
defined.

* platform/graphics/GLContext.cpp:
(WebCore::initializeOpenGLShimsIfNeeded): In case of GTK and WPE
initially we are going to keep TextureMapper using system GL
libraries, even when we are using ANGLE for WebGL, so we need the
shims definitions.
* platform/graphics/nicosia/texmap/NicosiaGCGLLayer.cpp:
(Nicosia::GCGLLayer::swapBuffersIfNeeded): We need to allocate and
update the layer buffer in the compositor thread to make sure it
is handled in just one thread.


Canonical link: https://commits.webkit.org/245195@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286978 268f45cc-cd09-0410-ab3c-d52691b4dbfc