Null pointer crash in FetchResponse::clone https://bugs.webkit.org/show_bug.cgi?id=234236 <rdar://86327601>
Reviewed by Alex Christensen.

From the log, we are most probably getting a null globalObject from a ScriptExecutionContext in FetchResponse::clone.
This may happen in case the document is navigated away but we still execute some code for it.
Add a null check to ensure we do not crash.

* Modules/fetch/FetchResponse.cpp:
(WebCore::FetchResponse::clone):



Canonical link: https://commits.webkit.org/245222@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@287017 268f45cc-cd09-0410-ab3c-d52691b4dbfc