Automatically generate event handler content attribute maps so they are easier to maintain https://bugs.webkit.org/show_bug.cgi?id=234254
Reviewed by Alexey Shvayka.

Source/WebCore:

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateHeader): Declare forEachEventHandlerContentAttribute
and forEachWindowEventHandlerContentAttribute as needed.
(GenerateImplementation): Call GenerateForEachEventHandlerContentAttribute
as needed.
(GenerateForEachEventHandlerContentAttribute): Define
forEachEventHandlerContentAttribute or forEachWindowEventHandlerContentAttribute.

* bindings/scripts/IDLAttributes.json: Added
GenerateForEachEventHandlerContentAttribute and
GenerateForEachWindowEventHandlerContentAttribute. Even though each
is used in only one class, it's still better to not hard-code the class names.

* html/HTMLAttributeNames.in: Added oncuechange, onrejectionhandled, and
onunhandledrejection even though we are not yet implementing any of them in
this behavior-preserving patch. They need to be present so the automatically
generated code compiles. But added explicit exceptions as mentioned below to
make sure no changes are implemented in this patch, and will take those exceptions
out later in a patch intended to produce a progression, likely with new test cases.
Re-sorted alphabetically, rather an arbitrary semi-sorted order.

* html/HTMLBodyElement.cpp:
(WebCore::HTMLBodyElement::createWindowEventHandlerNameMap): Deleted.
(WebCore::HTMLBodyElement::eventNameForWindowEventHandlerAttribute): Use
JSHTMLBodyElement::forEachWindowEventHandlerContentAttribute to create the map.
Use a lambda instead of a separate function and take advantage of the
deduction guide for NeverDestroyed. Added exceptions for onrejectionhandled
and onunhandledrejection, which are both in the IDL file but were not handled
here, along with a FIXME suggesting we remove the exceptions.
* html/HTMLElement.h: Removed createWindowEventHandlerNameMap.
* html/HTMLBodyElement.idl: Added GenerateForEachWindowEventHandlerContentAttribute.

* html/HTMLElement.cpp:
(WebCore::HTMLElement::createEventHandlerNameMap): Deleted.
(WebCore::HTMLElement::populateEventHandlerNameMap): Deleted.
(WebCore::HTMLElement::eventNameForEventHandlerAttribute): Use
JSHTMLElement::forEachEventHandlerContentAttribute to create the map.
Use a lambda instead of a separate function and take advantage of the
deduction guide for NeverDestroyed. Added exception for oncuechange,
which is in the IDL file but was not handled here, and for 19 other attributes,
which are not in the IDL file but were handled here, along with two FIXMEs
suggesting we remove the exceptions.
* html/HTMLElement.h: Removed populateEventHandlerNameMap and
createEventHandlerNameMap.
* html/HTMLElement.idl: Added GenerateForEachEventHandlerContentAttribute.

Source/WTF:

* wtf/NeverDestroyed.h: Add a deduction guide for NeverDestroyed.
In the future we can use this instead of makeNeverDestroyed or writing
out the entire template class type including template arguments.



Canonical link: https://commits.webkit.org/245224@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@287019 268f45cc-cd09-0410-ab3c-d52691b4dbfc